requirements.txt
fastapi
uvicorn
supabase-py
python-dotenv
pydantic